update to 6.0.1
Changed

Breaking: libxdot fields for the size and number of operations, the
statistics counts, and polygon line points are now size_t values instead of
int values
Accuracy of the bounding boxes printed by the showboxes feature have been
improved.

Fixed

Id attribute is not used in linearGradient. #2258

Graphviz 5.0.1 undocumented change of automatically generated output filename
with -O flag (missing dot separator). This was a regression in 5.0.1. #2270

Assert fail in aaglex for multiple calls to agmemread. This was a
regression in 5.0.1. #2272

Removed

The $GV_FILE_PATH sandboxing mechanism has been removed #2257

Signed-off-by: Zach Bacon <zachbacon@vba-m.com>